requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
